Split mm lora default test to separate file
